+++ compliance/profiles/full_enroll/controls/chef360-services.rb
@@ -53,11 +53,11 @@ control 'chef360-courier-log-check' do
   impact 1.0
   title 'Verify courier-runner log contains expected data'
   desc 'Check that the courier-runner log contains the string "failed to dequeue"'
-  
+
   # Set log path based on OS
   log_path = os.windows? ? 'C:\\hab\\svc\\courier-runner\\logs\\courier-log' : '/hab/svc/courier-runner/logs/courier-log'
-  
-  only_if("Courier log file exists") do
+
+  only_if('Courier log file exists') do
     file(log_path).exist?
   end