+++ compliance/profiles/partial_enroll/controls/chef360-services.rb
@@ -24,16 +24,16 @@ control 'chef360-courier-log-check' do
   impact 1.0
   title 'Verify courier-runner log contains expected data'
   desc 'Check that the courier-runner log contains the string "failed to dequeue"'
-  
+
   # Skip this control on Windows systems
   only_if('Not running on Windows') do
     !os.windows?
   end
-  
+
   # Set log path
   log_path = '/etc/chef_platform/tools/courier-runner/logs/courier-log'
-  
-  only_if("Courier log file exists") do
+
+  only_if('Courier log file exists') do
     file(log_path).exist?
   end